web/client: extract useEntityMap from AppRoutes
